+++ framework/validators/RequiredValidator.php
@@ -97,7 +97,7 @@ class RequiredValidator extends Validator
     /**
      * @inheritdoc
      */
-    protected function getClientOptions($model, $attribute)
+    public function getClientOptions($model, $attribute)
     {
         $options = [];
         if ($this->requiredValue !== null) {